Extract account and validate CLI commands
